Custom child of geonode.maps..Map moved to mapstory.mapstories..Map
